Restructure traits/enums to own mods & from/to raw
