+++ src/prelude.rs
@@ -1,4 +1,8 @@
 //! Default and recommended imports for functionality of crate.
 
-pub use series::{SeriesTrait, Series, SeriesEnumRef};
-pub use dataframe::{DataFrame, ColumnManager};
+pub use traits::BlackJackData;
+pub use series::{Series};
+pub use series::traits::SeriesTrait;
+pub use enums::{SeriesEnumRef, DType};
+pub use dataframe::{DataFrame};
+pub use dataframe::traits::{ColumnManager};
